remove redundant long casting

+++ SeperateBlackAndWhiteBalls/src/Main.java
@@ -12,7 +12,7 @@ public class Main {
             for (int i = 0; i < s.length(); i++) {
                 if (s.charAt(i) == '0') {
                     // Swap wti all black balls to the left
-                    countSwaps += (long) blackBallCount;
+                    countSwaps += blackBallCount;
                 } else {
                     //increment black ball count
                     blackBallCount++;
